# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.EmailVerificationResponse:
    description: Returned when email is verified and account activated
    properties:
      message:
        example: Email verified and Keycloak account created.
        type: string
      username:
        example: johndoe
        type: string
      verified:
        example: true
        type: boolean
    type: object
...
  handlers.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGc...
        type: string
    type: object
  handlers.RefreshResponse:
    properties:
      access_token:
        example: eyJhbGc...
        type: string
      expires_in:
        example: 300
        type: integer
      refresh_token:
        example: eyJhbGc...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      username:
        example: johndoe
        type: string
    type: object
  handlers.RegisterResponse:
    properties:
      message:
        example: user registered successfully, please verify email
        type: string
    type: object
...
  handlers.UnlockStatusResponse:
    properties:
      branding_locked:
        example: true
        type: boolean
      expires_at:
        example: "2025-08-14T02:50:19Z"
        type: string
      secured_by:
        example: "Peitho \U0001F510"
        type: string
      server_time:
        example: "2025-05-16T09:30:00Z"
        type: string
      trace_engine_active:
        example: true
        type: boolean
      unlocked:
        example: true
        type: boolean
      unlocked_at:
        example: "2025-05-16T02:50:19Z"
        type: string
    type: object
  handlers.UnlockSuccessResponse:
    description: Response for successful license unlock
    properties:
      message:
        example: unlocked
        type: string
    type: object
...
  observer.Actor:
    enum:
    - DEV
    - USER
    - HACKER
    type: string
    x-enum-varnames:
    - DEV
    - USER
    - HACKER
  observer.ThreatLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-comments:
      LEVEL_LOCKDOWN: 4 - System-level tampering, triggers lockdown
      LEVEL_MEAN: 3 - Mid-tier severity, often for hacks
      LEVEL_MOCK: 2 - Snarky, non-critical
      LEVEL_NONE: "0"
      LEVEL_SOFT: 1 - Minor mistake, gentle pushback
    x-enum-varnames:
    - LEVEL_NONE
    - LEVEL_SOFT
    - LEVEL_MOCK
    - LEVEL_MEAN
    - LEVEL_LOCKDOWN
  observer.TracePayload:
    properties:
      actor:
        allOf:
        - $ref: '#/definitions/observer.Actor'
        description: Who triggered it
        example: HACKER
      event:
        description: Event type
        example: tamper
        type: string
      id:
        description: Unique trace ID
        example: abc123
        type: string
      lock:
        description: Lockdown trigger flag
        example: false
        type: boolean
      message:
        description: Event message
        example: unauthorized access detected
        type: string
      severity:
        allOf:
        - $ref: '#/definitions/observer.ThreatLevel'
        description: ThreatLevel (as int for frontend)
        example: 3
      timestamp:
        description: UTC time
        example: "2025-05-16T12:30:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Secure backend API built with Go, Keycloak, and SQLite.
  title: PeithoSecure Lite API
  version: "1.0"
paths:
...
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Uses a valid refresh token to issue a new access token
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RefreshResponse'
        "400":
          description: Malformed request or missing refresh token
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a local user, sends verification email, and defers password/keycloak
        setup
      parameters:
      - description: Email and username payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RegisterResponse'
        "400":
          description: Malformed request or missing fields
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "409":
          description: Email already registered and verified
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Database or email error
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Register a new user (deferred setup)
      tags:
      - auth
...
  /api/v1/auth/resend-token:
    get:
      description: Sends a new verification email if the account exists and is not
        verified
      parameters:
      - description: User's registered email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing or already verified
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Email sending or DB error
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Resend email verification link
      tags:
      - auth
...
  /api/v1/auth/send-verification:
    post:
      consumes:
      - application/json
      description: Sends a new verification email to the user
      parameters:
      - description: Email payload
        in: body
        name: emailRequest
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - text/plain
      responses:
        "200":
          description: Verification email sent
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      summary: Resend verification email
      tags:
      - Email
...
  /api/v1/auth/unlock-status:
    get:
      description: Returns license unlock state, branding status, and server clock
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UnlockStatusResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: License Unlock Status
      tags:
      - License
  /api/v1/auth/unlock/validate:
    post:
      consumes:
      - application/json
      description: Accepts a signed PQC license block, writes it to disk, and validates
        it
      parameters:
      - description: License block payload
        in: body
        name: unlockRequest
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Unlock succeeded
          schema:
            $ref: '#/definitions/handlers.UnlockSuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "403":
          description: Unlock validation failed
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Write failed
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Validate unlock license
      tags:
      - license
  /api/v1/auth/verify-email:
    get:
      description: Verifies email via token, marks user as verified, registers in
        Keycloak, and issues PQC license
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EmailVerificationResponse'
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "409":
          description: Already verified
          schema:
            $ref: '#/definitions/observer.TracePayload'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/observer.TracePayload'
      summary: Verify user email
      tags:
      - Email
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"